plugin framework: update_command refactored as command
